Add the remainder of POSIX signals (plus WINCH and IO)
I did a quick check against OS X and OpenBSD. SIGPOLL is supposed to be part of
POSIX.1-2008, with the note that it might be removed in the future, and isn't
in either of those OSes, so removed that. SIGWINCH and SIGIO seem pretty
popular, so did include those (they are in the rust libc crate).

Refs:
http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/signal.h.html
http://man.openbsd.org/sigaction.2
https://developer.apple.com/legacy/library/documentation/Darwin/Reference/ManPages/man3/signal.3.html